linux-qcom-next: add v6.15 kernel recipe
Add a recipe for the linux-qcom-next 6.15 kernel.

This recipe compiles the qcom-next kernel tree, which includes additional
patches intended for upstream submission on top of the mainline kernel.

Starting with version 6.15-rc6, this tree will continue to track the latest
kernel releases (e.g., 6.15-rc7, 6.16, etc.), with patches being dropped or
rebased as needed. The eventual goal is to upstream all patches.

The qcom-next kernel tree will coexist with the linux-qcom-staging tree for
a few weeks, until it catches up with all required features. Post that, the
linux-qcom-staging tree will be deprecated and the corresponding recipe will
be removed.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>